Added custom integer field to the model.
Field named "custom_1" added. Not used on default but has been added because
there is the need for additional fields in other applications. So I added a
first custom data field which can be used in other applications.

Please note that you need to overwrite the table and form configuration to add
a form element to set the value for the new field.